# lib/openapi.yaml
openapi: 3.0.3
info:
  title: SPICE SGP4 API
  description: |
    REST API for SGP4 satellite orbit propagation using NASA/JPL NAIF CSPICE compiled to WebAssembly.

    ## Features
    - Parse Two-Line Element (TLE) sets into orbital elements
    - Parse Orbital Mean-Elements Message (OMM) JSON into orbital elements
    - Convert between TLE and OMM formats
    - Propagate satellite state to any epoch
    - Convert between UTC and Ephemeris Time
    - Support for WGS-72 and WGS-84 geophysical models

    ## Data Formats
    - **TLE**: Legacy Two-Line Element format (69 characters per line)
    - **OMM**: CCSDS Orbital Mean-Elements Message (JSON format, CCSDS 502.0-B-2)

    ## Reference Frames
    All position and velocity outputs are in the TEME (True Equator Mean Equinox) reference frame,
    which is the standard frame for SGP4 propagation.
  version: 1.0.0
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  contact:
    name: GitHub Repository
    url: https://github.com/haisamido/spice
servers:
  - url: http://localhost:50000
    description: Local development server
...
paths:
...
  /api/spice/sgp4/parse:
    post:
      summary: Parse TLE
      description: |
        Parse a Two-Line Element (TLE) set into orbital elements.
        Returns the epoch and 10-element array used by the SGP4 propagator.
      tags:
        - SPICE/SGP4/TLE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TLEInput'
            example:
              line1: "1 25544U 98067A   24015.50000000  .00016717  00000-0  10270-3 0  9025"
              line2: "2 25544  51.6400 208.9163 0006703  30.0825 330.0579 15.49560830    19"
      responses:
        '200':
          description: Successfully parsed TLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResponse'
        '400':
          description: Missing or invalid TLE lines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/propagate:
    post:
      summary: Propagate TLE or OMM
      description: |
        Unified propagation endpoint supporting both TLE and OMM input formats.

        **Single time mode**: Provide only `t0` to get one state vector at that time.

        **Range mode**: Provide `t0`, `tf`, and `step` to get an array of state vectors
        from t0 to tf at each step interval. Maximum of 1,209,602 points allowed per request (14 days at 1-second resolution).

        **Input format**: Use `input_type=tle` (default) for TLE input or `input_type=omm` for OMM input.
      tags:
        - SPICE/SGP4
      parameters:
        - name: t0
          in: query
          required: true
          description: Start time / single time (UTC string, e.g., "2024-01-15T12:00:00")
          schema:
            type: string
        - name: tf
          in: query
          required: false
          description: End time (UTC string). Required for range mode.
          schema:
            type: string
        - name: step
          in: query
          required: false
          description: Time step size. Required for range mode.
          schema:
            type: number
        - name: unit
          in: query
          description: Time step unit (only used in range mode)
          schema:
            type: string
            enum: [sec, min]
            default: sec
        - name: wgs
          in: query
          description: Geophysical model to use
          schema:
            type: string
            enum: [wgs72, wgs84]
            default: wgs72
        - name: input_type
          in: query
          description: Input format type
          schema:
            type: string
            enum: [tle, omm]
            default: tle
        - name: output_type
          in: query
          description: Output format (json or txt). TXT format returns CSV with datetime,et,x,y,z,vx,vy,vz columns.
          schema:
            type: string
            enum: [json, txt]
            default: txt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/TLEInput'
                - $ref: '#/components/schemas/OMMPropagateInput'
            examples:
              tle:
                summary: TLE input
                value:
                  line1: "1 25544U 98067A   24015.50000000  .00016717  00000-0  10270-3 0  9025"
                  line2: "2 25544  51.6400 208.9163 0006703  30.0825 330.0579 15.49560830    19"
              omm:
                summary: OMM input
                value:
                  omm:
                    OBJECT_NAME: "ISS (ZARYA)"
                    OBJECT_ID: "1998-067A"
                    EPOCH: "2024-01-15T12:00:00.000"
                    MEAN_MOTION: 15.49560830
                    ECCENTRICITY: 0.0006703
                    INCLINATION: 51.6400
                    RA_OF_ASC_NODE: 208.9163
                    ARG_OF_PERICENTER: 30.0825
                    MEAN_ANOMALY: 330.0579
                    NORAD_CAT_ID: 25544
                    BSTAR: 0.00010270
                    MEAN_MOTION_DOT: 0.00016717
      responses:
        '200':
          description: Successfully propagated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropagateRangeResponse'
            text/csv:
              schema:
                type: string
              example: |
                datetime,et,x,y,z,vx,vy,vz
                2024-01-15T12:00:00.000,758592069.18,-5945.93,-3284.80,0.29,2.31,-4.16,6.01
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/omm/parse:
    post:
      summary: Parse OMM
      description: |
        Parse an Orbital Mean-Elements Message (OMM) JSON into orbital elements.
        Also returns the equivalent TLE representation.
      tags:
        - SPICE/SGP4/OMM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OMMInput'
            example:
              OBJECT_NAME: "ISS (ZARYA)"
              OBJECT_ID: "1998-067A"
              EPOCH: "2024-01-15T12:00:00.000"
              MEAN_MOTION: 15.49560830
              ECCENTRICITY: 0.0006703
              INCLINATION: 51.6400
              RA_OF_ASC_NODE: 208.9163
              ARG_OF_PERICENTER: 30.0825
              MEAN_ANOMALY: 330.0579
              NORAD_CAT_ID: 25544
              BSTAR: 0.00010270
              MEAN_MOTION_DOT: 0.00016717
      responses:
        '200':
          description: Successfully parsed OMM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OMMParseResponse'
        '400':
          description: Missing or invalid OMM fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/omm/to-tle:
    post:
      summary: Convert OMM to TLE
      description: Convert an Orbital Mean-Elements Message (OMM) JSON to Two-Line Element format.
      tags:
        - SPICE/SGP4/OMM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OMMInput'
            example:
              OBJECT_NAME: "ISS (ZARYA)"
              OBJECT_ID: "1998-067A"
              EPOCH: "2024-01-15T12:00:00.000"
              MEAN_MOTION: 15.49560830
              ECCENTRICITY: 0.0006703
              INCLINATION: 51.6400
              RA_OF_ASC_NODE: 208.9163
              ARG_OF_PERICENTER: 30.0825
              MEAN_ANOMALY: 330.0579
              NORAD_CAT_ID: 25544
              BSTAR: 0.00010270
              MEAN_MOTION_DOT: 0.00016717
      responses:
        '200':
          description: Successfully converted to TLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TLEOutput'
        '400':
          description: Missing or invalid OMM fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/tle/to-omm:
    post:
      summary: Convert TLE to OMM
      description: Convert a Two-Line Element set to Orbital Mean-Elements Message (OMM) JSON format.
      tags:
        - SPICE/SGP4/TLE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TLEInput'
            example:
              line1: "1 25544U 98067A   24015.50000000  .00016717  00000-0  10270-3 0  9025"
              line2: "2 25544  51.6400 208.9163 0006703  30.0825 330.0579 15.49560830    19"
              name: "ISS (ZARYA)"
      responses:
        '200':
          description: Successfully converted to OMM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OMMOutput'
        '400':
          description: Missing or invalid TLE lines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/time/utc-to-et:
    get:
      summary: Convert UTC to Ephemeris Time
      description: |
        Convert a UTC time string to Ephemeris Time (ET).
        ET is measured in seconds past J2000 TDB (Barycentric Dynamical Time).
      tags:
        - SPICE/SGP4/Time
      parameters:
        - name: utc
          in: query
          required: true
          description: UTC time string (e.g., "2024-01-15T12:00:00")
          schema:
            type: string
          example: "2024-01-15T12:00:00"
      responses:
        '200':
          description: Successfully converted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UTCtoETResponse'
        '400':
          description: Missing or invalid UTC parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/spice/sgp4/time/et-to-utc:
    get:
      summary: Convert Ephemeris Time to UTC
      description: |
        Convert Ephemeris Time (ET) to a UTC time string.
        ET is measured in seconds past J2000 TDB.
      tags:
        - SPICE/SGP4/Time
      parameters:
        - name: et
          in: query
          required: true
          description: Ephemeris time in seconds past J2000 TDB
          schema:
            type: number
          example: 758894469.184
      responses:
        '200':
          description: Successfully converted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ETtoUTCResponse'
        '400':
          description: Missing or invalid ET parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/models/wgs/{name}:
    get:
      summary: Get model details
      description: Get detailed information about a specific WGS geophysical model.
      tags:
        - Models
      parameters:
        - name: name
          in: path
          required: true
          description: Model name
          schema:
            type: string
            enum: [wgs72, wgs84]
          example: wgs72
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeophysicalModel'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    TLEInput:
      type: object
      required:
        - line1
        - line2
      properties:
        line1:
          type: string
          description: TLE line 1 (69 characters)
        line2:
          type: string
          description: TLE line 2 (69 characters)
        name:
          type: string
          description: Satellite name (optional, used for TLE to OMM conversion)
...
    Position:
      type: array
      items:
        type: number
      minItems: 3
      maxItems: 3
      description: Cartesian position [x, y, z] in TEME frame (km)
      example: [-5945.93, -3284.80, 0.29]
...
    ParseResponse:
      type: object
      properties:
        epoch:
          type: number
          description: Epoch as ephemeris time (seconds past J2000 TDB)
        elements:
          type: array
          items:
            type: number
          description: 10-element orbital elements array
      example:
        epoch: 758592069.184
        elements: [0.0006703, 51.64, 208.9163, 30.0825, 330.0579, 15.4956083, 0.00016717, 0.0, 0.0001027, 0.0]
...
    PropagateRangeResponse:
      type: object
      properties:
        states:
          type: array
          items:
            type: object
            properties:
              et:
                type: number
                description: Ephemeris time (seconds past J2000 TDB)
              datetime:
                type: string
                description: UTC calendar representation (ISO 8601)
              position:
                $ref: '#/components/schemas/Position'
              velocity:
                $ref: '#/components/schemas/Velocity'
        epoch:
          type: number
          description: TLE epoch (seconds past J2000 TDB)
        model:
          type: string
          description: Geophysical model used
        count:
          type: integer
          description: Number of state vectors returned
        t0:
          type: string
          description: Start time
        tf:
          type: string
          description: End time
        step:
          type: number
          description: Time step
        unit:
          type: string
          description: Time step unit (sec or min)
        input_type:
          type: string
          description: Input format used (tle or omm)
          enum: [tle, omm]
      example:
        states:
          - datetime: "2024-01-15T12:00:00.000"
            et: 758592069.18
            position: [-5945.93, -3284.80, 0.29]
            velocity: [2.31, -4.16, 6.01]
        epoch: 758592069.18
        model: "wgs72"
        count: 1
        t0: "2024-01-15T12:00:00"
        input_type: "tle"
    UTCtoETResponse:
      type: object
      properties:
        et:
          type: number
          description: Ephemeris time (seconds past J2000 TDB)
      example:
        et: 758592069.184
    ETtoUTCResponse:
      type: object
      properties:
        utc:
          type: string
          description: UTC time string
      example:
        utc: "2024-01-15T12:00:00.000"
...
    GeophysicalModel:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        source:
          type: string
        constants:
          $ref: '#/components/schemas/GeophysicalConstants'
        units:
          type: object
          additionalProperties:
            type: string
      example:
        name: "WGS-72"
        description: "World Geodetic System 1972 constants for SGP4"
        source: "Vallado, Revisiting Spacetrack Report #3"
        constants:
          J2: 0.001082616
          J3: -0.00000253881
          J4: -0.00000165597
          KE: 0.0743669161
          QO: 120.0
          SO: 78.0
          RE: 6378.135
          AE: 1.0
        units:
          J2: "dimensionless"
          J3: "dimensionless"
          J4: "dimensionless"
          KE: "earth-radii^(3/2) / minute"
          QO: "km"
          SO: "km"
          RE: "km"
          AE: "distance units / Earth radius"
    OMMInput:
      type: object
      required:
        - OBJECT_NAME
        - OBJECT_ID
        - EPOCH
        - MEAN_MOTION
        - ECCENTRICITY
        - INCLINATION
        - RA_OF_ASC_NODE
        - ARG_OF_PERICENTER
        - MEAN_ANOMALY
        - NORAD_CAT_ID
        - BSTAR
        - MEAN_MOTION_DOT
      properties:
        CCSDS_OMM_VERS:
          type: string
          description: OMM version (default "2.0")
        CREATION_DATE:
          type: string
          description: Creation timestamp
        ORIGINATOR:
          type: string
          description: Data originator
        OBJECT_NAME:
          type: string
          description: Satellite name
        OBJECT_ID:
          type: string
          description: International designator (e.g., "1998-067A")
        CENTER_NAME:
          type: string
          description: Central body (default "EARTH")
        REF_FRAME:
          type: string
          description: Reference frame (default "TEME")
        TIME_SYSTEM:
          type: string
          description: Time system (default "UTC")
        MEAN_ELEMENT_THEORY:
          type: string
          description: Propagation theory (default "SGP4")
        EPOCH:
          type: string
          description: Epoch timestamp (ISO 8601)
        MEAN_MOTION:
          type: number
          description: Mean motion (revolutions/day)
        ECCENTRICITY:
          type: number
          description: Orbital eccentricity
        INCLINATION:
          type: number
          description: Inclination (degrees)
        RA_OF_ASC_NODE:
          type: number
          description: Right ascension of ascending node (degrees)
        ARG_OF_PERICENTER:
          type: number
          description: Argument of pericenter (degrees)
        MEAN_ANOMALY:
          type: number
          description: Mean anomaly (degrees)
        EPHEMERIS_TYPE:
          type: integer
          description: Ephemeris type (default 0 for SGP4)
        CLASSIFICATION_TYPE:
          type: string
          description: Classification (U=Unclassified, C=Classified, S=Secret)
        NORAD_CAT_ID:
          type: integer
          description: NORAD catalog number
        ELEMENT_SET_NO:
          type: integer
          description: Element set number
        REV_AT_EPOCH:
          type: integer
          description: Revolution number at epoch
        BSTAR:
          type: number
          description: Drag coefficient (1/Earth radii)
        MEAN_MOTION_DOT:
          type: number
          description: First derivative of mean motion (rev/day^2)
        MEAN_MOTION_DDOT:
          type: number
          description: Second derivative of mean motion (rev/day^3)
        SEMI_MAJOR_AXIS:
          type: number
          description: Semi-major axis in km (alternative to MEAN_MOTION)
        GM:
          type: number
          description: Gravitational parameter (km^3/s^2)
        REF_FRAME_EPOCH:
          type: string
          description: Reference frame epoch (ISO 8601)
        SPACECRAFT:
          $ref: '#/components/schemas/OMMSpacecraftParams'
        COVARIANCE:
          $ref: '#/components/schemas/OMMCovariance'
        USER_DEFINED:
          type: object
          additionalProperties: true
          description: User-defined key-value pairs
...
    OMMPropagateInput:
      type: object
      required:
        - omm
      properties:
        omm:
          $ref: '#/components/schemas/OMMInput'
    OMMParseResponse:
      type: object
      properties:
        epoch:
          type: number
          description: Epoch as ephemeris time (seconds past J2000 TDB)
        elements:
          type: array
          items:
            type: number
          description: 10-element orbital elements array
        tle:
          type: object
          properties:
            line1:
              type: string
            line2:
              type: string
      example:
        epoch: 758592069.184
        elements: [0.0006703, 51.64, 208.9163, 30.0825, 330.0579, 15.4956083, 0.00016717, 0.0, 0.0001027, 0.0]
        tle:
          line1: "1 25544U 98067A   24015.50000000  .00016717  00000-0  10270-3 0  9025"
          line2: "2 25544  51.6400 208.9163 0006703  30.0825 330.0579 15.49560830    19"
    OMMOutput:
      $ref: '#/components/schemas/OMMInput'
    TLEOutput:
      type: object
      properties:
        line1:
          type: string
          description: TLE line 1 (69 characters)
        line2:
          type: string
          description: TLE line 2 (69 characters)
        name:
          type: string
          description: Satellite name (optional)
      example:
        line1: "1 25544U 98067A   24015.50000000  .00016717  00000-0  10270-3 0  9025"
        line2: "2 25544  51.6400 208.9163 0006703  30.0825 330.0579 15.49560830    19"
        name: "ISS (ZARYA)"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message